add dashboard container
