+++ src/components/index.ts
@@ -1,4 +1,6 @@
-export { default as ControlledFlow } from "./ui/ControlledFlow"
 export { default as SignupFlow } from "./containers/SignupFlow"
+export { default as Dashboard } from "./containers/Dashboard"
+
+export { default as ControlledFlow } from "./ui/ControlledFlow"
 export { default as Button } from "./ui/Button"
 export { default as SplitScreen } from "./ui/SplitScreen"
